Adding half of the map and polishing some parts of the code!
Half of the map has been done! For the sake of saving some space and hard coding the map, I've added the "static_coordinates" array that gets different values depending if we want to make the right or the left side of the map! Also when calculating the coordinates from left to right, I had to make another if(), because at the right side we remove from the value of the "left-to-right" coordinate, while at the left side of the map, we add to the value of the "left-to-right" coordinate. The way I did it is a quick solution. I will try to make it better in the future!