Making the animations
The way we make the animations is using the same setInterval we used for making the movement but this time we use setTimeout too to make the second of the images appear after the first one. We use the .classList method which can add classes to our <div> tag  through javascript! That way we use the "id" of the <div> element to move the character and use the "class" of the <div> element to add pictures (and changing them of course to look like an animation) to it!