"Score" system
I've added a scoring system in which when pacman moves through a tile, a new property added to his object called "score" goes by 10 if there is a coin in it! Also pacman can't collect coins he's already collected! The only thing left in this scoring and collecting system is making the coins disappear when pacman goes through them!